requirements.txt
python-telegram-bot==20.7
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
python-dateutil==2.8.2
Pillow==10.2.0
aiohttp==3.9.3
nest-asyncio==1.6.0
python-dotenv==1.0.0